requirements.txt
python-dotenv
websocket-client
requests